# optimizer_openapi.json
{"openapi":"3.1.0","info":{"title":"GridKey BESS Optimizer API","description":"Energy optimization API with Weather, Price, Battery, and Optimizer services","version":"1.0.0"},"servers":[{"url":"https://superindulgent-francine-nonpessimistic.ngrok-free.dev","description":"ngrok tunnel"}],"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/weather/forecast":{"get":{"tags":["Weather"],"summary":"Get Weather Forecast","description":"Get weather-based renewable energy generation forecast.\n\nReturns hourly PV and Wind generation predictions for the specified location.","operationId":"get_weather_forecast_weather_forecast_get","parameters":[{"name":"location","in":"query","required":false,"schema":{"type":"string","description":"City name (Munich, Berlin, Shanghai) or coordinates","default":"Munich","title":"Location"},"description":"City name (Munich, Berlin, Shanghai) or coordinates"},{"name":"hours","in":"query","required":false,"schema":{"type":"integer","maximum":120,"minimum":1,"description":"Forecast hours (1-120)","default":24,"title":"Hours"},"description":"Forecast hours (1-120)"},{"name":"pv_capacity_kw","in":"query","required":false,"schema":{"type":"number","description":"PV capacity in kW","default":10.0,"title":"Pv Capacity Kw"},"description":"PV capacity in kW"},{"name":"wind_capacity_kw","in":"query","required":false,"schema":{"type":"number","description":"Wind turbine capacity in kW","default":50.0,"title":"Wind Capacity Kw"},"description":"Wind turbine capacity in kW"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherForecastResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/price/forecast":{"get":{"tags":["Price"],"summary":"Get Price Forecast","description":"Get electricity market prices for all markets.\n\nReturns Day-Ahead, FCR, aFRR Capacity, and aFRR Energy prices for the specified country.","operationId":"get_price_forecast_price_forecast_get","parameters":[{"name":"country","in":"query","required":false,"schema":{"type":"string","description":"Country/Bidding zone (DE_LU, AT, CH, HU, CZ)","default":"DE_LU","title":"Country"},"description":"Country/Bidding zone (DE_LU, AT, CH, HU, CZ)"},{"name":"hours","in":"query","required":false,"schema":{"type":"integer","maximum":168,"minimum":1,"description":"Forecast hours (1-168)","default":48,"title":"Hours"},"description":"Forecast hours (1-168)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/optimize":{"post":{"tags":["Optimizer"],"summary":"Flexible Horizon Optimization","description":"optimize_flexible\n\nPure logic optimization. \nAccepts market prices and renewable generation.\nReturns optimal schedule.","operationId":"optimize_flexible_api_v1_optimize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/optimize-mpc":{"post":{"tags":["Optimizer"],"summary":"MPC Rolling Horizon (12h)","description":"optimize_mpc\n\nFixed 12h horizon using Rolling Horizon strategy.\nRequires 48 data points (15-min resolution).","operationId":"optimize_mpc_api_v1_optimize_mpc_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"GenerationPointResponse":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"pv_output_kw":{"type":"number","title":"Pv Output Kw"},"wind_output_kw":{"type":"number","title":"Wind Output Kw"},"total_output_kw":{"type":"number","title":"Total Output Kw"}},"type":"object","required":["timestamp","pv_output_kw","wind_output_kw","total_output_kw"],"title":"GenerationPointResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MarketPrices":{"properties":{"day_ahead":{"items":{"type":"number"},"type":"array","title":"Day Ahead","description":"DA Prices (EUR/MWh)"},"afrr_energy_pos":{"items":{"type":"number"},"type":"array","title":"Afrr Energy Pos","description":"aFRR+ Energy Prices (EUR/MWh)"},"afrr_energy_neg":{"items":{"type":"number"},"type":"array","title":"Afrr Energy Neg","description":"aFRR- Energy Prices (EUR/MWh)"},"fcr":{"items":{"type":"number"},"type":"array","title":"Fcr","description":"FCR Capacity Prices (EUR/MW)"},"afrr_capacity_pos":{"items":{"type":"number"},"type":"array","title":"Afrr Capacity Pos","description":"aFRR+ Capacity Prices (EUR/MW)"},"afrr_capacity_neg":{"items":{"type":"number"},"type":"array","title":"Afrr Capacity Neg","description":"aFRR- Capacity Prices (EUR/MW)"}},"type":"object","required":["day_ahead","afrr_energy_pos","afrr_energy_neg","fcr","afrr_capacity_pos","afrr_capacity_neg"],"title":"MarketPrices","description":"Market prices arrays.\nStrictly enforcing array structure as per API Guide."},"OptimizationData":{"properties":{"objective_value":{"type":"number","title":"Objective Value"},"net_profit":{"type":"number","title":"Net Profit"},"revenue_breakdown":{"additionalProperties":{"type":"number"},"type":"object","title":"Revenue Breakdown"},"degradation_cost":{"type":"number","title":"Degradation Cost"},"cyclic_aging_cost":{"type":"number","title":"Cyclic Aging Cost"},"calendar_aging_cost":{"type":"number","title":"Calendar Aging Cost"},"schedule":{"items":{"$ref":"#/components/schemas/ScheduleEntry"},"type":"array","title":"Schedule"},"soc_trajectory":{"items":{"type":"number"},"type":"array","title":"Soc Trajectory"},"solve_time_seconds":{"type":"number","title":"Solve Time Seconds"},"solver_name":{"type":"string","title":"Solver Name"},"model_type":{"type":"string","title":"Model Type"},"status":{"type":"string","title":"Status"},"renewable_utilization":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Renewable Utilization"}},"type":"object","required":["objective_value","net_profit","revenue_breakdown","degradation_cost","cyclic_aging_cost","calendar_aging_cost","schedule","soc_trajectory","solve_time_seconds","solver_name","model_type","status"],"title":"OptimizationData"},"OptimizeRequest":{"properties":{"location":{"type":"string","title":"Location","example":"Munich"},"country":{"type":"string","title":"Country","example":"DE_LU"},"model_type":{"type":"string","title":"Model Type","description":"Solver Model Type","default":"III-renew"},"c_rate":{"type":"number","maximum":2.0,"exclusiveMinimum":0.0,"title":"C Rate","default":0.5},"alpha":{"type":"number","title":"Alpha","description":"Degradation weight","default":1.0},"market_prices":{"$ref":"#/components/schemas/MarketPrices"},"renewable_generation":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Renewable Generation","description":"PV/Wind generation (kW)"},"time_horizon_hours":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Time Horizon Hours","description":"Optimization horizon in hours"}},"type":"object","required":["location","country","market_prices"],"title":"OptimizeRequest","description":"Pure API Request Body.\nNo hidden logic. Input-driven."},"OptimizeResponse":{"properties":{"status":{"type":"string","title":"Status"},"data":{"$ref":"#/components/schemas/OptimizationData"}},"type":"object","required":["status","data"],"title":"OptimizeResponse"},"ScheduleEntry":{"properties":{"timestamp":{"type":"string","title":"Timestamp"},"action":{"type":"string","title":"Action"},"power_kw":{"type":"number","title":"Power Kw"},"market":{"type":"string","title":"Market"},"soc_after":{"type":"number","title":"Soc After"},"renewable_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Renewable Action"},"renewable_power_kw":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Renewable Power Kw"}},"type":"object","required":["timestamp","action","power_kw","market","soc_after"],"title":"ScheduleEntry"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WeatherForecastResponse":{"properties":{"location":{"type":"string","title":"Location"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"},"timeline":{"items":{"$ref":"#/components/schemas/GenerationPointResponse"},"type":"array","title":"Timeline"}},"type":"object","required":["location","generated_at","timeline"],"title":"WeatherForecastResponse"}}}}